% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConnectionDetails.R
\name{createConnectionDetails}
\alias{createConnectionDetails}
\title{Create Connection Details for Databricks}
\usage{
createConnectionDetails(keyringName = Sys.getenv("keyringName"))
}
\arguments{
\item{keyringName}{A character string specifying the name of the keyring. Defaults to the value of the environment variable \code{keyringName}.}
}
\value{
A list containing the connection details.
}
\description{
This function creates connection details for a Databricks database using the DatabaseConnector package.
It retrieves the connection string, username, and password from a specified keyring.
}
