% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelExecution.R
\docType{package}
\name{ParallelExecution-package}
\alias{ParallelExecution}
\alias{ParallelExecution-package}
\title{ParallelExecution: An R package to run queries in parallel.}
\description{
An R package to run queries in parallel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/ParallelExecution/}
  \item \url{https://github.com/OHDSI/ParallelExecution}
  \item Report bugs at \url{https://github.com/OHDSI/ParallelExecution/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
